+++ compilation/src/main/java/cz/vutbr/fit/group/verifit/oslc/compilation/VeriFitCompilationManager.java
@@ -351,6 +351,34 @@ public class VeriFitCompilationManager {
     	else // (count > 1)
     		throw new OslcResourceException("Too many source parameters. Expected exactly one.");
     }
+    
+    /**
+	 * Check that the build command is defined if the compilation parameter is set to true
+	 * @param autoRequest			Automation Request with parameters to check
+	 * @throws OslcResourceException 	When the compilation parameter is true and the build command is null
+	 */
+    public static void checkSutDeployCompilationAndBuildParams (AutomationRequest autoRequest) throws OslcResourceException
+    {
+		// count the "source.*" input params
+		Boolean compileSet = true;
+		Boolean buildCmdFound = false;
+    	for (ParameterInstance submittedParam : autoRequest.getInputParameter())
+		{				
+			if (submittedParam.getName().equals("compile"))
+			{
+				compileSet = Boolean.valueOf(submittedParam.getValue());
+			}
+			else if (submittedParam.getName().equals("buildCommand"))
+			{
+				buildCmdFound = true;
+			}
+		}
+    	
+    	if (compileSet == true && buildCmdFound == false)
+    		throw new OslcResourceException("compilation is enabled, but the buildCommand is missing.");
+    	else
+    		return;
+    }
     // End of user code
 
     public static void contextInitializeServletListener(final ServletContextEvent servletContextEvent)
@@ -743,7 +771,10 @@ public class VeriFitCompilationManager {
 			// check that the request contains exactly one "source.*" parameter (can not be checked automatically based on the AutoPlan
 			// throws an exception if the Inputs are not OK
 			checkSutDeploySourceInputs(newResource);
-
+			
+			// check that buildCommand is not null, if compilation was set to true
+			checkSutDeployCompilationAndBuildParams(newResource);
+			
 			// create an AutomationResult for this AutoRequest
 			AutomationResult newAutoResult = createAutomationResultForAutomationRequest(newResource, outputParams);
 			newResource.setProducedAutomationResult(new Link(newAutoResult.getAbout()));