+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added a new parameter to the test script (-l to not launch the adapter)
 - renamed maven project model to lyo_designer_model
 - typos, refactor, ...
+- added a "compile" command to the compilation adapter that controls whether to perform compilation or not
 
 ### v1.3.1
 powershell modifications